# backend/05_APIs_Externas/api_rest/openapi.yaml
openapi: 3.0.3
info:
  title: METGO 3D API REST
  description: API agrometeorológica MVP — Valle de Aconcagua (Chile)
  version: 1.0.0-mvp
servers:
  - url: http://127.0.0.1:8080
    description: Desarrollo local
  - url: https://metgo-api.onrender.com
    description: Producción Render
...
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: Login JWT (E9 — claim sitio)
      description: |
        Emite JWT con `sub`, `role`, `tenant` (legacy) y **`sitio`** (producto).
        El body puede incluir `sitio` para scoped login; un usuario de Quillota
        no puede pedir `mantos_blancos` (403). Admin global (`sitio=null` en membresía)
        puede omitir sitio (acceso a todos) o pedirlo para limitar el token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, example: admin }
                password: { type: string, example: admin123 }
                sitio: { type: string, example: quillota, description: "Producto METGO (opcional)" }
      responses:
        "200":
          description: Token emitido (user.sitio en payload)
        "401":
          description: Credenciales incorrectas
        "403":
          description: Sitio no autorizado para el usuario
...
  /api/sitios:
    get:
      tags: [meteo]
      summary: Listar sitios METGO
      description: Catálogo multi-producto (quillota, paine, demo, …).
      security: [{ bearerAuth: [] }]
      parameters:
        - name: incluir_plantilla
          in: query
          schema: { type: string, enum: ["0", "1"], default: "1" }
      responses:
        "200":
          description: Lista de sitios (slug, nombre, dominio, primary, modules, …)
        "401":
          description: Sin token
  /api/estaciones:
    get:
      tags: [meteo]
      summary: Listar estaciones
      description: |
        Por defecto `sitio=quillota` (Valle de Aconcagua).
        Otros: `paine` (Torres del Paine), `copiapo` (calidad del aire), `demo` (plantilla E6).
      security: [{ bearerAuth: [] }]
      parameters:
        - name: sitio
          in: query
          required: false
          schema:
            type: string
            enum: [quillota, paine, copiapo, demo]
            default: quillota
          description: Sitio METGO multi-producto
      responses:
        "200":
          description: Lista de estaciones (id, nombre, lat, lon, sitio, …)
        "401":
          description: Sin token
  /api/public/sitios:
    get:
      tags: [meteo]
      summary: Listar sitios (público)
      parameters:
        - name: incluir_plantilla
          in: query
          schema: { type: string, enum: ["0", "1"], default: "1" }
      responses:
        "200":
          description: Lista de sitios
  /api/public/estaciones:
    get:
      tags: [meteo]
      summary: Listar estaciones (público)
      parameters:
        - name: sitio
          in: query
          required: false
          schema:
            type: string
            enum: [quillota, paine, copiapo, demo]
            default: quillota
      responses:
        "200":
          description: Lista de estaciones
  /api/public/aire/{estacion_id}:
    get:
      tags: [aire]
      summary: Calidad del aire actual (público)
      description: |
        Open-Meteo Air Quality (modelo CAMS) + **ICAP** chileno calculado server-side
        (contaminante rector PM2.5/PM10, categorías Bueno…Emergencia) y
        recomendaciones de salud por categoría. Sitio E7 Copiapó.
      parameters:
        - name: estacion_id
          in: path
          required: true
          schema: { type: string, example: copiapo_centro }
      responses:
        "200":
          description: PM2.5, PM10, SO2, NO2, O3, CO, polvo, icap, nivel, recomendaciones
        "404":
          description: Estación desconocida
        "503":
          description: Servicio CAMS no disponible
...
  /api/public/meteo/{estacion_id}:
    get:
      tags: [meteo]
      summary: Resumen meteo público
      parameters:
        - name: estacion_id
          in: path
          required: true
          schema: { type: string, example: base_torres }
      responses:
        "200":
          description: Resumen del día
        "404":
          description: Estación desconocida
        "503":
          description: OpenMeteo no disponible
  /api/public/meteo/{estacion_id}/pronostico:
    get:
      tags: [meteo]
      summary: Pronóstico diario público
      parameters:
        - name: estacion_id
          in: path
          required: true
          schema: { type: string, example: base_torres }
        - name: dias
          in: query
          schema: { type: integer, default: 7 }
      responses:
        "200":
          description: Lista de días
        "404":
          description: Estación desconocida
        "503":
          description: OpenMeteo no disponible
  /api/meteo/{estacion_id}:
    get:
      tags: [meteo]
      summary: Resumen meteorológico
      security: [{ bearerAuth: [] }]
      parameters:
        - name: estacion_id
          in: path
          required: true
          schema: { type: string, example: quillota }
        - name: tipo
          in: query
          schema: { type: string, enum: [pronostico, historico] }
      responses:
        "200":
          description: Datos meteo
        "401":
          description: Sin token
        "404":
          description: Sin datos
        "503":
          description: Servicio de OpenMeteo temporalmente no disponible
  /api/meteo/{estacion_id}/pronostico:
    get:
      tags: [meteo]
      summary: Pronóstico diario (OpenMeteo persistido en Supabase; fallback desde BD)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: estacion_id
          in: path
          required: true
          schema: { type: string, example: quillota }
        - name: dias
          in: query
          schema: { type: integer, default: 7, minimum: 1, maximum: 16 }
      responses:
        "200":
          description: Lista diaria; `desde_cache=true` si se sirvió desde caché/BD
        "401":
          description: Sin token
        "404":
          description: Estación no encontrada
        "503":
          description: OpenMeteo no disponible y sin datos persistidos
  /api/meteo/{estacion_id}/historico:
    get:
      tags: [meteo]
      summary: Histórico diario (OpenMeteo + registros persistidos en Supabase)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: estacion_id
          in: path
          required: true
          schema: { type: string, example: quillota }
        - name: dias
          in: query
          schema: { type: integer, default: 30, minimum: 1, maximum: 92 }
      responses:
        "200":
          description: Lista diaria deduplicada (reales sobre sintéticos)
        "401":
          description: Sin token
        "503":
          description: OpenMeteo no disponible y sin datos persistidos
  /api/meteo/{estacion_id}/viento-horario:
    get:
      tags: [meteo]
      summary: Rosa de vientos horaria (dirección+velocidad) OpenMeteo forecast
      security: [{ bearerAuth: [] }]
      parameters:
        - name: estacion_id
          in: path
          required: true
          schema: { type: string, example: quillota }
        - name: dias
          in: query
          schema: { type: integer, default: 7, minimum: 1, maximum: 16 }
      responses:
        "200":
          description: Arrays paralelos `direcciones` y `velocidades` (unidad m/s)
        "404":
          description: Sin datos de viento
        "503":
          description: Servicio de OpenMeteo temporalmente no disponible
  /api/alertas:
    get:
      tags: [sistema]
      summary: Alertas activas (automáticas + configuradas)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: estacion
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Lista de alertas
...
  /api/meteo/comparativo:
    get:
      tags: [meteo]
      summary: Comparativo multi-estación
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Resumen por estación principal
  /api/meteo/comparativo/historico:
    get:
      tags: [meteo]
      summary: Histórico multi-estación (valle)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: dias
          in: query
          schema: { type: integer, default: 14, maximum: 92 }
      responses:
        "200":
          description: Series diarias por estación principal
  /api/metricas/globales:
    get:
      tags: [sistema]
      summary: KPIs globales del valle
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Métricas consolidadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  estaciones_activas: { type: integer }
                  temperatura_media_max: { type: number, nullable: true }
                  temperatura_media_min: { type: number, nullable: true }
                  precipitacion_total: { type: number }
                  viento_max: { type: number }
                  humedad_media: { type: number }
                  alertas_activas: { type: integer }
                  alertas_warning: { type: integer }
                  referencia_fecha: { type: string, format: date, description: Día calendario Chile }
                  detalle_estaciones:
                    type: array
                    items:
                      type: object
                      properties:
                        estacion_id: { type: string }
                        estacion: { type: string }
                        temperatura: { type: number }
                        temperatura_max: { type: number }
                        temperatura_min: { type: number }
                        humedad: { type: number }
                        viento: { type: number }
                        precipitacion: { type: number }
                        fuente: { type: string }
                  estaciones: { type: array, items: { type: string } }
                  actualizado: { type: string, format: date-time }
...
  /api/catalogo:
    get:
      tags: [sistema]
      summary: Catálogo de módulos (alias de modulos)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: categoria
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Módulos del sistema
  /api/modulos:
    get:
      tags: [sistema]
      summary: Catálogo de módulos
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Módulos
  /api/servicios/streamlit/{modulo_id}/visor:
    get:
      tags: [servicios]
      summary: URL del visor de puertos (iframe)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: modulo_id
          in: path
          required: true
          schema: { type: string, example: visualizaciones }
      responses:
        "200":
          description: URLs embed local/nube
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT